\name{generateSim}
\alias{generateSim}
\title{Generate simulations according several factors...}
\usage{generateSim(template, factors)
}
\description{Generate simulations according several factors}
\value{A list which contain all simulations. Row names of factors 
are used for simulation names.}
\arguments{\item{template}{File path to template sim files}
\item{factors}{A data frame which contained all combinations of factors.
The column names of data frame are parameter names in the sim file.}
}

