\name{left}
\alias{left}
\title{Get the several characters of a string from left...}
\usage{left(str, num=1)
}
\description{Get the several characters of a string from left}
\value{the character vector of the string from left}
\arguments{\item{str}{Input string}
\item{num}{The number of character will be returned from left}
}

