\name{right}
\alias{right}
\title{Get the several characters of a string from right...}
\usage{right(str, num=1)
}
\description{Get the several characters of a string from right}
\value{the character vector of the string from right}
\arguments{\item{str}{Input string}
\item{num}{The number of character will be returned from right}
}

