\name{searchChar}
\alias{searchChar}
\title{Search a character from a string...}
\usage{searchChar(str, search, start=1)
}
\description{Search a character from a string}
\value{The postion where character firstly appear}
\arguments{\item{str}{Input string}
\item{search}{The character will be searched}
\item{start}{The start position to search}
}

