\name{vector2string}
\alias{vector2string}
\title{Convert vector to string...}
\usage{vector2string(vector, sep=", ")
}
\description{Convert vector to string}
\details{Convert vertor to string with a separater}
\value{The string which connected all members of vector}
\arguments{\item{vector}{The input vector to convert}
\item{sep}{The separater with default value ", "}
}

